requirements.txt
flask
zhupuai
python-dotenv
langchain_community
langchain